Move AutoValueExtension to its own package and merge the extension_api build target into value_processor. Remove some duplicated or obsolete code resulting from merges. Apply various other source cleanups. ------------- Created by MOE: https://github.com/google/moe MOE_MIGRATED_REVID=101855945
